✨ feat(config.ts): add gitpush and prefix configuration options
The config.ts file now supports two new configuration options: gitpush and prefix. The gitpush option can be set to false to disable automatic pushing to git after committing. The prefix option can be set to a string or a regular expression to add a prefix to the commit message generated by opencommit.

🐛 fix(utils/git.ts): add getCurrentGitBranch function

The getCurrentGitBranch function has been added to the utils/git.ts file. This function returns the current git branch name. This is used in the prepare-commit-msg-hook.ts file to generate the prefix from the branch name when using a regular expression as the prefix configuration option.